Fix focus loss when setting fullscreen
If this extension is running, when a window changes to fullscreen
mode, it loses the focus. This is very inconvenient when playing a
movie: when the user sets fullscreen, (s)he has to do an extra click
on the movie to give to the player the focus, and be able to control
it from the keyboard.

This patch fixes it.